Add files via upload
Deep Learning and Reinforcement Learning
Final Project Deliverable


Section 2: Brief Description of the Dataset
Source: The diabetes dataset used in this analysis was one of the datasets used in cousera class.
Description: The dataset contains 65% of a benign class (non-cancerous) and 35% of a malignant class (Cancerous).  This we consider to be imbalanced. The instances are described by 9 attributes, all of which are numerical (int and float). They include: